Rename header folder to enet6
