+++ include/enet6/protocol.h
@@ -5,7 +5,7 @@
 #ifndef __ENET_PROTOCOL_H__
 #define __ENET_PROTOCOL_H__
 
-#include "enet/types.h"
+#include "enet6/types.h"
 
 enum
 {
