+++ include/enet6/utility.h
