+++ src/callbacks.c
@@ -3,7 +3,7 @@
  @brief ENet callback functions
 */
 #define ENET_BUILDING_LIB 1
-#include "enet/enet.h"
+#include "enet6/enet.h"
 
 static ENetCallbacks callbacks = { malloc, free, abort };
