+++ src/packet.c
@@ -4,7 +4,7 @@
 */
 #include <string.h>
 #define ENET_BUILDING_LIB 1
-#include "enet/enet.h"
+#include "enet6/enet.h"
 
 /** @defgroup Packet ENet packet functions 
     @{
